Backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
pymongo
nltk
pydantic[email]